requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
pytesseract
Pillow
langchain
openai
faiss-cpu
tiktoken
python-dotenv
python-jose
sqlalchemy
alembic
passlib[bcrypt]
langchain-community
langchain-openai
python-multipart